requirements.txt
numpy
pandas
Levenshtein
sklearn.cluster
pathlib
tkinter